test: migrate ResultView tests in source
